requirements.txt
atlassian-python-api
pymongo